+++ MagicMuscle/Assets/Animations/EnemyDeath.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4a6844c2454f8491880c88c236fc29a2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
