+++ MagicMuscle/Assets/New Folder.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b83a7879da4d74d1db88baef564226ba
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
