+++ MagicMuscle/Assets/New Folder/Enemy.controller.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 646ef60a3d51242b18d45ec7e93fe089
+guid: ba299ad8cce464275a451ad449649fea
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
