+++ MagicMuscle/Assets/Prefabs/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ebcb0d42a660f427f965948f92bd36b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
