+++ MagicMuscle/Assets/Prefabs/Humberger.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 59d3c7619b91644389e87d7292e59c8b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
